+++ jax/scipy/linalg.py
@@ -22,6 +22,7 @@ from jax._src.scipy.linalg import (
   eigh_tridiagonal as eigh_tridiagonal,
   expm as expm,
   expm_frechet as expm_frechet,
+  hessenberg as hessenberg,
   inv as inv,
   lu as lu,
   lu_factor as lu_factor,